+++ src/cjs/Polygon.d.ts
@@ -27,7 +27,8 @@
  * @modified 2023-09-25 Added the `Polygon.getInterpolationPolygon(number)` function.
  * @modified 2023-09-25 Added the `Polygon.lineIntersections(Line,boolean)` function.
  * @modified 2023-09-29 Added the `Polygon.closestLineIntersection(Line,boolean)` function.
- * @version 1.11.0
+ * @modified 2023-11-24 Added the `Polygon.containsPolygon(Polygon)' function.
+ * @version 1.12.0
  *
  * @file Polygon
  * @public
@@ -141,6 +142,17 @@ export declare class Polygon implements SVGSerializable {
      * @memberof Polygon
      **/
     containsVert(vert: XYCoords): boolean;
+    /**
+     * Check if the passed polygon is completly contained inside this polygon.
+     *
+     * This means:
+     *  - all polygon's vertices must be inside this polygon.
+     *  - the polygon has no edge intersections with this polygon.
+     *
+     * @param polygon
+     * @return {boolean}
+     */
+    containsPolygon(polygon: Polygon): boolean;
     /**
      * Calculate the area of the given polygon (unsigned).
      *